Solution architect design

+++ src/Core/ProcurementTracker.Domain/Entities/PurchaseRequest.cs
@@ -2,11 +2,11 @@
 {
     public class PurchaseRequest : BaseAuditableEntity
     {
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public PurchaseRequestStatus PurchaseRequestStatus { get; set; }
         public DateTime RequiredDeliveryDate { get; set; }
         public string Description { get; set; }
-        public int SupplierId { get; set; }
+        public long SupplierId { get; set; }
 
 
         public virtual User StatusChangedBy { get; set; }
